New Big Weight gag icon (Thanks, Jerky!)
I've come to the realisation that I'm not too keen on the overcomplicated process of redoing the actual gag's texture to match the icon, so I'm flipping the process by replacing the gag icon instead.

It's a little less invasive as it prevents replacing textures in areas that don't need it, improving the ability of mixing and match this pack with other content packs. It also slightly shrinks the size of the pack, which is one of my focuses - I don't want this pack to be bigger than it needs to be.

Special thanks to Jerky for allowing me to use his Big Weight gag icon in my pack! You can find him on Twitter here: https://twitter.com/ToonJerky