Halved map texture resolution
In TTO, it was 256x256. In base TTR, it's 1024x1024. I feel like that's a little too much detail for something that is ultimately not gonna be seen that often, so I've chosen to bring it down to 512x512. The result is very nice and if I never spoke of this change you probably wouldn't even notice.

Original: https://cdn.discordapp.com/attachments/1014963084126846998/1096106505742340158/1024x.png
New: https://cdn.discordapp.com/attachments/1014963084126846998/1096106533370204240/512x.png