Remastered SpeedChat bubble
I decided to downscale the SpeedChat bubble texture as I felt it was way too high-resolution for such a core part of the gameplay. I gave it slightly thicker borders in an attempt to merge the feel of both TTR and TTO's SpeedChat bubbles, and the rounded corners now fill in more area.

Before: https://cdn.discordapp.com/attachments/1014963084126846998/1096154061641887905/old.png
After: https://cdn.discordapp.com/attachments/1014963084126846998/1096154070500253767/new.png

I may attempt to further edit this in order to improve the triangular part of the bubble, as it has gone untouched.

Note: I have not touched the Thought bubble texture due to the complexity of its design and inability to know the full texture space I can work with. It was much easier to determine this with the SpeedChat bubble texture. These do not rely on RGB files.